# src/CreatorPantry.Tests/Api/Snapshots/openapi-v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "CreatorPantry API",
    "description": "Browser clients call these routes through CreatorPantry.Gateway with a session cookie. Errors are RFC 9457 problem details carrying a stable \u0060code\u0060 and a \u0060traceId\u0060.",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/workspaces/{workspaceSlug}/recipes/{recipeId}/adaptation-requests/{requestId}": {
      "get": {
        "tags": [
          "RecipeAdaptationRequests"
        ],
        "summary": "Where a requested adaptation has got to, and the proposed changes once there are any.",
        "description": "Polling before the changes exist is a \u0060200\u0060 carrying a status, not a \u0060404\u0060 \u2014 the request\nresource exists from the moment it was accepted.",
        "parameters": [
          {
            "name": "workspaceSlug",
            "in": "path",
            "description": "Where a requested adaptation has got to, and the proposed changes once there are any.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "recipeId",
            "in": "path",
            "description": "Where a requested adaptation has got to, and the proposed changes once there are any.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "requestId",
            "in": "path",
            "description": "The id the request returned. It is the operation\u0027s id.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AiProposalStatus"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiProposalStatus"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiProposalStatus"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "default": {
            "description": "Problem details with a stable \u0060code\u0060 and \u0060traceId\u0060.",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bffSession": []
          }
        ]
      }
    },
...
    "/api/v1/workspaces/{workspaceSlug}/recipe-concept-requests/{requestId}": {
      "get": {
        "tags": [
          "RecipeConceptRequests"
        ],
        "summary": "Where a requested generation has got to, and the proposed concepts once there are any.",
        "description": "Polling before the concepts exist is a \u0060200\u0060 carrying a status, not a \u0060404\u0060 -- the request\nresource exists from the moment it was accepted, and reporting it missing would make a normal poll look\nlike a failure.",
        "parameters": [
          {
            "name": "workspaceSlug",
            "in": "path",
            "description": "Where a requested generation has got to, and the proposed concepts once there are any.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "requestId",
            "in": "path",
            "description": "The id the request returned. It is the operation\u0027s id.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AiProposalStatus"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiProposalStatus"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiProposalStatus"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "default": {
            "description": "Problem details with a stable \u0060code\u0060 and \u0060traceId\u0060.",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bffSession": []
          }
        ]
      }
    },
...
    "/api/v1/workspaces/{workspaceSlug}/recipes/{recipeId}": {
      "get": {
        "tags": [
          "Recipes"
        ],
        "summary": "Reads one recipe of the workspace named by the route.",
        "parameters": [
          {
            "name": "workspaceSlug",
            "in": "path",
            "description": "Bound only so the route is well formed. The workspace is resolved server-side from this segment and\nthe caller\u0027s membership before the action runs, and nothing here reads it (tenancy.md).",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "recipeId",
            "in": "path",
            "description": "The recipe to read. Constrained to a Guid, so a malformed id never reaches this action: routing answers\n404, the same status as an unknown recipe and as one belonging to another workspace. The problem body\ndiffers \u2014 an edge 404 carries the generic \u0060not_found\u0060 code rather than this module\u0027s \u2014 so nothing\nis disclosed, but a client branching on \u0060code\u0060 sees two codes for what is one condition to it.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeDetail"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "default": {
            "description": "Problem details with a stable \u0060code\u0060 and \u0060traceId\u0060.",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bffSession": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Recipes"
        ],
        "summary": "Changes part of one recipe and returns it as it now stands.",
        "description": "This is a JSON Merge Patch, not a replacement. A field the body does not mention is left exactly as\nit is; a field sent with a value is set to it; a field sent as \u0060null\u0060 is cleared. Omitting a field\nnever blanks it. \u0060tags\u0060 is the one field that replaces rather than merges \u2014 a submitted list becomes\nthe recipe\u0027s complete set, and \u0060[]\u0060 or \u0060null\u0060 removes them all. \u0060title\u0060 and \u0060status\u0060 may be changed\nbut not cleared. \u0060expectedConcurrencyToken\u0060 is required and is not a field of the recipe: it is the\n\u0060concurrencyToken\u0060 from the read this edit was composed against, and an edit quoting a token the\nrecipe has moved past is refused with \u0060409 recipes.recipe.conflict\u0060 rather than overwriting whoever\nsaved first. \u0060reason\u0060 is likewise not a recipe field; it is recorded on the version this edit writes.\nAn edit that would change nothing writes no version and leaves the token valid. The response is the\nwhole recipe, in the same shape a read returns, so an editor can rebind from it: it carries the\nrefreshed token the next edit must quote, and the version this one wrote.",
        "parameters": [
          {
            "name": "workspaceSlug",
            "in": "path",
            "description": "Bound only so the route is well formed. The workspace is resolved server-side from this segment and\nthe caller\u0027s membership before the action runs, and nothing here reads it (tenancy.md).",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "recipeId",
            "in": "path",
            "description": "The recipe to change. Constrained to a Guid, so a malformed id answers 404 at routing.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Optional. A repeat of the same key and the same edit returns the original response with\n\u0060Idempotent-Replayed: true\u0060 rather than being answered as a conflict.",
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/AntiforgeryToken"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRecipe"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRecipe"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRecipe"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeDetail"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeDetail"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "default": {
            "description": "Problem details with a stable \u0060code\u0060 and \u0060traceId\u0060.",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bffSession": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AiFailureCategory": {
        "enum": [
          "Unspecified",
          "Validation",
          "Quota",
          "TemplateUnavailable",
          "Provider",
          "RateLimited",
          "Timeout",
          "OutputSchemaInvalid",
          "DomainInvalid",
          "SafetyBlocked",
          "Cancelled",
          "LeaseAbandoned",
          "AccountSuspended"
        ],
        "type": "string"
      },
      "AiOperationScope": {
        "enum": [
          "Unspecified",
          "WholeRecipe",
          "Ingredients",
          "Instructions",
          "Metadata",
          "Media",
          "NotApplicable",
          "Advisory"
        ],
        "type": "string"
      },
      "AiOperationStatus": {
        "enum": [
          "Requested",
          "Running",
          "Proposed",
          "Accepted",
          "PartiallyAccepted",
          "Rejected",
          "Failed",
          "Expired"
        ],
        "type": "string"
      },
...
      "AiProposalStatus": {
        "required": [
          "aiProposalRequestId",
          "status",
          "taskType",
          "scope",
          "sourceVersionId",
          "requestedAt",
          "statusChangedAt",
          "failureCategory",
          "proposal"
        ],
        "type": "object",
        "properties": {
          "aiProposalRequestId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "$ref": "#/components/schemas/AiOperationStatus"
          },
          "taskType": {
            "$ref": "#/components/schemas/AiTaskType"
          },
          "scope": {
            "$ref": "#/components/schemas/AiOperationScope"
          },
          "sourceVersionId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "requestedAt": {
            "type": "string",
            "format": "date-time"
          },
          "statusChangedAt": {
            "type": "string",
            "format": "date-time"
          },
          "failureCategory": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AiFailureCategory"
              }
            ]
          },
          "proposal": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AiProposalDetail"
              }
            ]
          }
        }
      },
...
      "AiTaskType": {
        "enum": [
          "Unspecified",
          "Diagnostic",
          "RecipeConcepts",
          "RecipeFirstDraft",
          "RecipeRevision",
          "IngredientSubstitution",
          "RecipeAdaptation",
          "RecipeReview",
          "ProposalExplanation"
        ],
        "type": "string"
      },
...
      "ProblemDetails": {
        "required": [
          "code",
          "traceId"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "code": {
            "type": "string",
            "description": "Stable, machine-readable error code, e.g. \u0060auth.registration.invalid\u0060 or \u0060not_found\u0060."
          },
          "traceId": {
            "type": "string",
            "description": "Correlation id for support and tracing."
          }
        }
      },
...
      "RecipeDetail": {
        "required": [
          "id",
          "title",
          "description",
          "headnote",
          "notes",
          "storageNotes",
          "attributionText",
          "sourceUrl",
          "cuisineId",
          "courseId",
          "primaryTechniqueId",
          "prepTimeMinutes",
          "cookTimeMinutes",
          "restTimeMinutes",
          "totalTimeMinutes",
          "yieldText",
          "yieldQuantity",
          "yieldUnitId",
          "servingCount",
          "servingSize",
          "status",
          "duplicatedFrom",
          "createdAt",
          "updatedAt",
          "concurrencyToken",
          "currentVersion",
          "ingredientGroups",
          "instructionGroups",
          "equipment",
          "assetLinks",
          "tags"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "headnote": {
            "type": [
              "null",
              "string"
            ]
          },
          "notes": {
            "type": [
              "null",
              "string"
            ]
          },
          "storageNotes": {
            "type": [
              "null",
              "string"
            ]
          },
          "attributionText": {
            "type": [
              "null",
              "string"
            ]
          },
          "sourceUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "cuisineId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "courseId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "primaryTechniqueId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "prepTimeMinutes": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "cookTimeMinutes": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "restTimeMinutes": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalTimeMinutes": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "yieldText": {
            "type": [
              "null",
              "string"
            ]
          },
          "yieldQuantity": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d\u002B)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "yieldUnitId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "servingCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d\u002B)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "servingSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d\u002B)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "status": {
            "$ref": "#/components/schemas/RecipeStatus"
          },
          "duplicatedFrom": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RecipeDuplicateSource"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "concurrencyToken": {
            "type": "string"
          },
          "currentVersion": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RecipeVersionSummary"
              }
            ]
          },
          "ingredientGroups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecipeIngredientGroup"
            }
          },
          "instructionGroups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecipeInstructionGroup"
            }
          },
          "equipment": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecipeEquipment"
            }
          },
          "assetLinks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecipeAssetLink"
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecipeTag"
            }
          }
        }
      },
...
      "ValidationProblemDetails": {
        "required": [
          "code",
          "traceId"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "code": {
            "type": "string",
            "description": "Stable, machine-readable error code, e.g. \u0060auth.registration.invalid\u0060 or \u0060not_found\u0060."
          },
          "traceId": {
            "type": "string",
            "description": "Correlation id for support and tracing."
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "description": "Field errors keyed by camelCase request field name."
          }
        }
      },
...
    },
    "parameters": {
...
      "AntiforgeryToken": {
        "name": "X-XSRF-TOKEN",
        "in": "header",
        "description": "Antiforgery request token from GET /bff/antiforgery (or the sign-in response). Required by the gateway on every unsafe request; a missing or stale token is rejected with 400 csrf.invalid.",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
    },
...
    "securitySchemes": {
      "bffSession": {
        "type": "apiKey",
        "description": "HttpOnly session cookie issued by the gateway (BFF). Browsers never handle tokens. Provisional name, finalized with the BFF session.",
        "name": "__Host-creatorpantry-session",
        "in": "cookie"
      }
    }
  },
...
}